Remove temporary files.
