--- src/server/http2.js
@@ -1,80 +1,0 @@
-const http = require("http");
-const JsonRpcServerFactory = require("./index2");
-const HttpServerProtocol = require("./protocol/http");
-const { errorToStatus } = require("../constants");
-
-/**
- * Constructor for Jsonic HTTP server
- * @class HTTPServer
- * @constructor
- * @extends Client
- * @param {Object} [options] optional settings for server
- * @return HTTPServer
- */
-class HTTPServer extends JsonRpcServerFactory {
-  setServer() {
-    this.server = new http.Server();
-  }
-
-  buildProtocol() {
-    this.server.on("connection", (client) => {
-      this.connectedClients.push(client);
-      this.emit("clientConnected", client);
-      client.on("close", () => {
-        this.emit("clientDisconnected");
-      });
-      client.on("end", () => {
-        this.emit("clientDisconnected");
-      });
-    });
-    this.server.on("request", (request, response) => {
-      this.pcolInstance = new HttpServerProtocol(
-        this,
-        request,
-        response,
-        this.options.version,
-        this.options.delimiter
-      );
-      this.pcolInstance.clientConnected();
-    });
-  }
-
-  setResponseHeader({ response, errorCode, notification }) {
-    let statusCode = 200;
-    if (notification) {
-      statusCode = 204;
-    }
-    const header = {
-      "Content-Type": "application/json"
-    };
-    if (errorCode) {
-      statusCode = errorToStatus[String(errorCode)];
-    }
-    response.writeHead(statusCode, header);
-  }
-
-  clientConnected(cb) {
-    this.on("clientConnected", (client) => {
-      cb({
-        host: client.remoteAddress,
-        port: client.remotePort
-      });
-    });
-  }
-
-  clientDisconnected(cb) {
-    this.on("clientDisconnected", (client) => {
-      const clientIndex = this.connectedClients.findIndex((c) => client === c);
-      if (clientIndex === -1) {
-        return cb(`Unknown client ${JSON.stringify(client)}`);
-      }
-      const [deletedClient] = this.connectedClients.splice(clientIndex, 1);
-      return cb({
-        host: deletedClient.remoteAddress,
-        port: deletedClient.remotePort
-      });
-    });
-  }
-}
-
-module.exports = HTTPServer;
