+++ src/server/index.js
@@ -1,9 +1,8 @@
 const EventEmitter = require("events");
-const { formatResponse, formatError, BatchRequest } = require("../functions");
-const { ERR_CODES, ERR_MSGS } = require("../constants");
+const { formatResponse } = require("../functions");
 
 /**
- * @class Server
+ * @class JsonRpcServerFactory
  * @extends require('events').EventEmitter
  * @param {Object} [host] host IP to connect with
  * @param {Object} [host] host port to connect with
@@ -13,11 +12,11 @@ const { ERR_CODES, ERR_MSGS } = require("../constants");
  * @param {Boolean} [options.exlusive=false] disallow port sharing
  * @return {Client}
  */
-class Server extends EventEmitter {
+class JsonRpcServerFactory extends EventEmitter {
   constructor(options) {
     super();
-    if (!(this instanceof Server)) {
-      return new Server(options);
+    if (!(this instanceof JsonRpcServerFactory)) {
+      return new JsonRpcServerFactory(options);
     }
 
     const defaults = {
@@ -34,38 +33,47 @@ class Server extends EventEmitter {
     };
 
     this.methods = {};
+    this.connectedClients = [];
     this.listening = false;
+    this.pcolInstance = undefined;
   }
 
   listen() {
     return new Promise((resolve, reject) => {
       if (this.listening) {
         // not having this caused MaxEventListeners error
-        reject(Error("server already listening"));
+        return reject(Error("server already listening"));
       }
       const { host, port, exclusive } = this.options;
+      this.setServer();
       this.server.listen({ host, port, exclusive });
       this.server.on("listening", () => {
         this.listening = true;
-        try {
-          this.handleData();
-          this.handleError();
-          resolve({
-            host: this.server.address().address,
-            port: this.server.address().port
-          });
-        } catch (e) {
-          this.listening = false;
-          reject(e);
-        }
+        this.buildProtocol();
+        resolve({
+          host: this.server.address().address,
+          port: this.server.address().port
+        });
       });
-      this.server.on("error", (error) => {
-        this.listening = false;
-        reject(error);
-      });
-      this.server.on("close", () => {
-        this.listening = false;
-      });
+      this.setupListeners();
+    });
+  }
+
+  buildProtocol() {
+    throw new Error("function must be overwritten in subclass");
+  }
+
+  setupListeners() {
+    this.on("error", (error) => {
+      this.listening = false;
+      throw error;
+    });
+    this.server.on("error", (error) => {
+      this.listening = false;
+      throw error;
+    });
+    this.server.on("close", () => {
+      this.listening = false;
     });
   }
 
@@ -81,293 +89,106 @@ class Server extends EventEmitter {
     });
   }
 
+  // add the method and its associated callback to the object
   method(name, cb) {
-    // add the method and its associated callback to the object
     this.methods[name] = cb;
   }
 
   onNotify(method, cb) {
-    if (method === "clientConnected" || method === "clientDisconnected") {
-      throw new Error(`reserved event name ${method}`);
-    }
     this.on(method, cb);
   }
 
   removeOnNotify(method, cb) {
-    if (method === "clientConnected" || method === "clientDisconnected") {
-      throw new Error(`reserved event name ${method}`);
-    }
     this.removeListener(method, cb);
   }
 
   removeAllOnNotify(method) {
-    if (method === "clientConnected" || method === "clientDisconnected") {
-      throw new Error(`reserved event name ${method}`);
-    }
     this.removeAllListeners([method]);
   }
 
-  handleData() {
-    throw new Error("function must be overwritten in subsclass");
-  }
-
-  handleBatchRequest(requests) {
-    const batchRequests = requests
-      .map((request) => {
-        try {
-          const message = this.validateMessage(request);
-          if (message.notification) {
-            this.emit(message.notification.method, message.notification);
-            return;
-          }
-          return this.getResult(message)
-            .then((result) => JSON.parse(result))
-            .catch((error) => {
-              throw error;
-            });
-        } catch (e) {
-          return Promise.reject(e);
-        }
-      })
-      .filter((el) => el != null);
-    if (batchRequests.length === 0) {
-      return Promise.resolve({ empty: true });
+  notify(notifications) {
+    if (notifications.length === 0 || !Array.isArray(notifications)) {
+      throw new Error("Invalid arguments");
     }
-    return Promise.all(
-      batchRequests.map((promise) =>
-        promise.catch((error) => JSON.parse(error.message))
-      )
-    );
-  }
-
-  validateRequest(request) {
-    try {
-      const message = JSON.parse(request);
-      return message;
-    } catch (e) {
-      throw new Error(
-        formatError({
-          jsonrpc: this.options.version,
-          id: null,
-          code: ERR_CODES.parseError,
-          message: ERR_MSGS.parseError,
-          delimiter: this.options.delimiter
-        })
-      );
+    const responses = this._getNotificationResponses(notifications);
+    if (responses.length === 0) {
+      throw new Error("Unable to generate a response object");
     }
-  }
-
-  validateMessage(message) {
-    if (Array.isArray(message)) {
-      // possible batch request
-      if (message.length === 0) {
-        const error = formatError({
-          jsonrpc: this.options.version,
-          id: null,
-          code: ERR_CODES.invalidRequest,
-          message: ERR_MSGS.invalidRequest,
-          delimiter: this.options.delimiter
-        });
-        throw new Error(error);
-      }
-      throw new BatchRequest(undefined, message);
-    }
-
-    if (!(message === Object(message))) {
-      throw new Error(
-        formatError({
-          jsonrpc: this.options.version,
-          id: null,
-          code: ERR_CODES.invalidRequest,
-          message: ERR_MSGS.invalidRequest,
-          delimiter: this.options.delimiter
-        })
-      );
-    }
-
-    if (!(typeof message.method === "string")) {
-      throw new Error(
-        formatError({
-          jsonrpc: message.jsonrpc,
-          id: message.id,
-          code: ERR_CODES.invalidRequest,
-          message: ERR_MSGS.invalidRequest,
-          delimiter: this.options.delimiter
-        })
-      );
-    }
-
-    if (!message.id) {
-      // no id, so assume notification
-      return { notification: message };
-    }
-
-    if (message.jsonrpc) {
-      if (this.options.version !== "2.0") {
-        throw new Error(
-          formatError({
-            id: message.id,
-            code: ERR_CODES.invalidRequest,
-            message: ERR_MSGS.invalidRequest,
-            delimiter: this.options.delimiter
-          })
-        );
-      }
-    }
-
-    if (!(message.method in this.methods)) {
-      throw new Error(
-        formatError({
-          jsonrpc: message.jsonrpc,
-          id: message.id,
-          code: ERR_CODES.methodNotFound,
-          message: ERR_MSGS.methodNotFound,
-          delimiter: this.options.delimiter
-        })
-      );
-    }
-
-    if (
-      message.params &&
-      !Array.isArray(message.params) &&
-      !(message.params === Object(message.params))
-    ) {
-      throw new Error(
-        formatError({
-          jsonrpc: message.jsonrpc,
-          id: message.id,
-          code: ERR_CODES.invalidParams,
-          message: ERR_MSGS.invalidParams,
-          delimiter: this.options.delimiter
-        })
-      );
-    }
-    // data looks good
-    return message;
-  }
-
-  handleValidation(chunk) {
-    return new Promise((resolve, reject) => {
-      try {
-        const result = this.validateRequest(chunk);
-        const message = this.validateMessage(result);
-        resolve(message);
-      } catch (e) {
-        if (e instanceof BatchRequest) {
-          this.handleBatchRequest(e.request)
-            .then((result) => {
-              resolve({ batch: result });
-            })
-            .catch((error) => {
-              reject(error);
-            });
-        } else {
-          reject(e);
-        }
-      }
-    });
-  }
-
-  getResult(message) {
-    // function needs to be async since the method can be a promise
-    return new Promise((resolve, reject) => {
-      const { params } = message;
-      let error = formatError({
-        jsonrpc: message.jsonrpc,
-        id: message.id,
-        code: ERR_CODES.unknown,
-        message: ERR_MSGS.unknown,
-        delimiter: this.options.delimiter
+    let response;
+    if (responses.length === 1) {
+      response = formatResponse(responses[0]);
+    } else {
+      // batch notification responses
+      response = "[";
+      responses.forEach((res, idx) => {
+        response += formatResponse(res);
+        response += idx === responses.length - 1 ? "" : ",";
       });
-      try {
-        const result = params
-          ? this.methods[message.method](params)
-          : this.methods[message.method]();
-        if (
-          result &&
-          (typeof result.then === "function" || result instanceof Promise)
-        ) {
-          Promise.all([result])
-            .then((results) => {
-              resolve(
-                formatResponse({
-                  jsonrpc: message.jsonrpc,
-                  id: message.id,
-                  result: results || 0,
-                  delimiter: this.options.delimiter
-                })
-              );
-            })
-            .catch((resError) => {
-              error = formatError({
-                jsonrpc: message.jsonrpc,
-                id: message.id,
-                code: ERR_CODES.internal,
-                message: `${JSON.stringify(resError.message || resError)}`,
-                delimiter: this.options.delimiter
-              });
-              reject(error);
-            });
-        } else {
-          resolve(
-            formatResponse({
-              jsonrpc: message.jsonrpc,
-              id: message.id,
-              result: result || 0,
-              delimiter: this.options.delimiter
-            })
-          );
-        }
-      } catch (e) {
-        if (e instanceof TypeError) {
-          error = formatError({
-            jsonrpc: message.jsonrpc,
-            id: message.id,
-            code: ERR_CODES.invalidParams,
-            message: ERR_MSGS.invalidParams,
-            delimiter: this.options.delimiter
-          });
-        }
-        reject(error);
+      response += "]";
+    }
+    return this.sendNotifications(response);
+  }
+
+  _getNotificationResponses(notifications) {
+    return notifications.map(([method, params]) => {
+      if (!method && !params) {
+        throw new Error("Unable to generate a response object");
       }
+      const response = {
+        method,
+        params,
+        delimiter: this.options.delimiter
+      };
+      if (this.options.version === "2.0") {
+        response.jsonrpc = "2.0";
+      }
+      return response;
     });
   }
 
-  clientConnected() {
-    throw new Error("function must be overwritten in subclass");
+  clientConnected(cb) {
+    this.on("clientConnected", (client) => {
+      cb({
+        host: client.remoteAddress,
+        port: client.remotePort
+      });
+    });
   }
 
-  clientDisconnected() {
-    throw new Error("function must be overwritten in subsclass");
-  }
-
-  handleError() {
-    this.on("error", (error) => {
-      this.listening = false;
-      throw error;
+  clientDisconnected(cb) {
+    this.on("clientDisconnected", (client) => {
+      const clientIndex = this.connectedClients.findIndex((c) => client === c);
+      if (clientIndex === -1) {
+        return cb(`Unknown client ${JSON.stringify(client)}`);
+      }
+      const [deletedClient] = this.connectedClients.splice(clientIndex, 1);
+      return cb({
+        host: deletedClient.remoteAddress,
+        port: deletedClient.remotePort
+      });
     });
   }
 }
 
-module.exports = Server;
+module.exports = JsonRpcServerFactory;
 
 /**
  * HTTP server constructor
  * @type ServerHTTP
  * @static
  */
-Server.http = require("./http2");
+JsonRpcServerFactory.http = require("./http");
 
 /**
  * TCP server constructor
  * @type ServerTCP
  * @static
  */
-Server.tcp = require("./tcp2");
+JsonRpcServerFactory.tcp = require("./tcp");
 
 /**
  * WS server constructor
  * @type ServerWS
  * @static
  */
-Server.ws = require("./ws2");
+JsonRpcServerFactory.ws = require("./ws");
